requirements.txt
streamlit
google-cloud-storage
google-generativeai
PyPDF2
reportlab
